pyproject.toml
[tool.poetry]
name = "tasktracker"
version = "0.1.0"
description = ""
authors = ["nnnLik"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.12"
django = "^4.2.3"
djangorestframework = "^3.14.0"
djoser = "^2.2.0"
django-cors-headers = "^4.1.0"
drf-yasg = "^1.21.6"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.6"
django-filter = "^23.2"
pillow = "^10.0.0"
coreapi = "^2.3.3"
redis = "^4.6.0"
django-redis = "^5.3.0"
celery = "^5.3.1"
flower = "^2.0.0"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
